Integrate Typescript with Jest & Adjust lint-staged config
