+++ src/actions/__tests__/user.test.ts
@@ -1,10 +1,11 @@
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import axios from 'axios';
+// @ts-ignore
 import httpAdapter from 'axios/lib/adapters/http';
 import nock from 'nock';
 
-import { fetchUsers } from '../users';
+import { fetchUser } from '../user';
 
 const host = 'http://localhost';
 
@@ -13,42 +14,50 @@ axios.defaults.adapter = httpAdapter;
 
 const mockStore = configureMockStore([thunk]);
 
-describe('fetch users data', () => {
-  const response = [{ id: 'test', name: 'Welly' }];
+describe('fetch user data', () => {
+  const userId = 'test';
+  const response = {
+    name: 'Welly',
+    phone: '007',
+    email: 'test@gmail.com',
+    website: 'www.test.com'
+  };
   const errorMessage = 'Request failed with status code 404';
 
   afterEach(() => {
     nock.disableNetConnect();
   });
 
-  it('creates USERS_SUCCESS when fetching users has been done', () => {
+  it('creates USER_SUCCESS when fetching user has been done', () => {
     nock(host)
       .get('/test')
       .reply(200, response);
 
     const expectedActions = [
-      { type: 'USERS_REQUESTING' },
-      { type: 'USERS_SUCCESS', data: response }
+      { type: 'USER_REQUESTING', userId },
+      { type: 'USER_SUCCESS', userId, data: response }
     ];
-    const store = mockStore({ list: null });
+    const store = mockStore({ info: null });
 
-    store.dispatch(fetchUsers(`${host}/test`)).then(() => {
+    // @ts-ignore
+    store.dispatch(fetchUser('test', host)).then(() => {
       expect(store.getActions()).toEqual(expectedActions);
     });
   });
 
-  it('creates USERS_FAILURE when fail to fetch users', () => {
+  it('creates USER_FAILURE when fail to fetch user', () => {
     nock(host)
       .get('/test')
       .replyWithError(errorMessage);
 
     const expectedActions = [
-      { type: 'USERS_REQUESTING' },
-      { type: 'USERS_FAILURE', err: errorMessage }
+      { type: 'USER_REQUESTING', userId },
+      { type: 'USER_FAILURE', userId, err: errorMessage }
     ];
     const store = mockStore({ err: null });
 
-    store.dispatch(fetchUsers(`${host}/test`)).then(() => {
+    // @ts-ignore
+    store.dispatch(fetchUser('test', host)).then(() => {
       expect(store.getActions()).toEqual(expectedActions);
     });
   });
