+++ src/actions/__tests__/users.test.ts
@@ -1,10 +1,11 @@
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import axios from 'axios';
+// @ts-ignore
 import httpAdapter from 'axios/lib/adapters/http';
 import nock from 'nock';
 
-import { fetchUser } from '../user';
+import { fetchUsers } from '../users';
 
 const host = 'http://localhost';
 
@@ -13,48 +14,44 @@ axios.defaults.adapter = httpAdapter;
 
 const mockStore = configureMockStore([thunk]);
 
-describe('fetch user data', () => {
-  const userId = 'test';
-  const response = {
-    name: 'Welly',
-    phone: '007',
-    email: 'test@gmail.com',
-    website: 'www.test.com'
-  };
+describe('fetch users data', () => {
+  const response = [{ id: 'test', name: 'Welly' }];
   const errorMessage = 'Request failed with status code 404';
 
   afterEach(() => {
     nock.disableNetConnect();
   });
 
-  it('creates USER_SUCCESS when fetching user has been done', () => {
+  it('creates USERS_SUCCESS when fetching users has been done', () => {
     nock(host)
       .get('/test')
       .reply(200, response);
 
     const expectedActions = [
-      { type: 'USER_REQUESTING', userId },
-      { type: 'USER_SUCCESS', userId, data: response }
+      { type: 'USERS_REQUESTING' },
+      { type: 'USERS_SUCCESS', data: response }
     ];
-    const store = mockStore({ info: null });
+    const store = mockStore({ list: null });
 
-    store.dispatch(fetchUser('test', host)).then(() => {
+    // @ts-ignore
+    store.dispatch(fetchUsers(`${host}/test`)).then(() => {
       expect(store.getActions()).toEqual(expectedActions);
     });
   });
 
-  it('creates USER_FAILURE when fail to fetch user', () => {
+  it('creates USERS_FAILURE when fail to fetch users', () => {
     nock(host)
       .get('/test')
       .replyWithError(errorMessage);
 
     const expectedActions = [
-      { type: 'USER_REQUESTING', userId },
-      { type: 'USER_FAILURE', userId, err: errorMessage }
+      { type: 'USERS_REQUESTING' },
+      { type: 'USERS_FAILURE', err: errorMessage }
     ];
     const store = mockStore({ err: null });
 
-    store.dispatch(fetchUser('test', host)).then(() => {
+    // @ts-ignore
+    store.dispatch(fetchUsers(`${host}/test`)).then(() => {
       expect(store.getActions()).toEqual(expectedActions);
     });
   });
