+++ tools/jest/assetMock.ts
