+++ tools/jest/styleMock.ts
@@ -1,3 +1,4 @@
+// @ts-ignore
 import idObj from 'identity-obj-proxy';
 
 export default idObj;
